+++ src/services/panel/DiffPanel.ts
@@ -3,17 +3,19 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
 
-import { remove } from '../../@l13/natvies/arrays';
+import { remove } from '../../@l13/arrays';
+import { isMacOs, isWindows } from '../../@l13/platforms';
 import { Diff, DiffCopyMessage, DiffFile, DiffInitMessage, DiffMultiCopyMessage, StatsMap, Uri } from '../../types';
-import { isMacOs, isWindows } from '../@l13/nodes/platforms';
 import { formatNameAndDesc } from '../@l13/utils/formats';
+
+import { open } from '../../common/dialogs';
+import * as files from '../../common/files';
+import * as settings from '../../common/settings';
 
 import { DiffCompare } from '../actions/DiffCompare';
 import { DiffCopy } from '../actions/DiffCopy';
 import { DiffDelete } from '../actions/DiffDelete';
 import { DiffOpen } from '../actions/DiffOpen';
-import { DiffDialog } from '../common/DiffDialog';
-import { DiffSettings } from '../common/DiffSettings';
 import { DiffOutput } from '../output/DiffOutput';
 import { DiffResult } from '../output/DiffResult';
 import { DiffStats } from '../output/DiffStats';
@@ -255,13 +257,13 @@ export class DiffPanel {
 	//	open
 		
 		this.msg.on('open:diffToSide', (diff:Diff) => DiffOpen.open(diff, true));
-		this.msg.on('open:diff', (diff:Diff) => DiffOpen.open(diff, DiffSettings.get('openToSide', false)));
+		this.msg.on('open:diff', (diff:Diff) => DiffOpen.open(diff, settings.get('openToSide', false)));
 		
-		this.msg.on('reveal:file', (pathname:string) => DiffOpen.reveal(pathname));
+		this.msg.on('reveal:file', (pathname:string) => files.reveal(pathname));
 		
 		this.msg.on('open:dialog', async () => {
 			
-			const fsPath = await DiffDialog.open();
+			const fsPath = await open();
 			
 			this.msg.send('open:dialog', { fsPath });
 			
@@ -372,8 +374,9 @@ export class DiffPanel {
 	
 	private getHTMLforDiff (webview:vscode.Webview) {
 		
-		const scriptUri = vscode.Uri.file(path.join(this.context.extensionPath, 'media', 'main.js'));
-		const styleUri = vscode.Uri.file(path.join(this.context.extensionPath, 'media', 'style.css'));
+		const mediaPath = path.join(this.context.extensionPath, 'media');
+		const scriptUri = vscode.Uri.file(path.join(mediaPath, 'main.js'));
+		const styleUri = vscode.Uri.file(path.join(mediaPath, 'style.css'));
 		
 		const nonceToken = nonce();
 		const csp = `default-src 'none'; img-src ${webview.cspSource} data:; style-src 'unsafe-inline' ${webview.cspSource}; script-src 'nonce-${nonceToken}';`;
@@ -405,19 +408,21 @@ export class DiffPanel {
 	
 	public static async create (context:vscode.ExtensionContext, uris:null|Uri[]|vscode.Uri[] = null, compare?:boolean) {
 		
+		const mediaPath = path.join(context.extensionPath, 'media');
+		const iconsPath = path.join(mediaPath, 'icons');
 		const panel = vscode.window.createWebviewPanel(DiffPanel.viewType, 'Diff Folders', {
 			viewColumn: vscode.ViewColumn.Active,
 		}, {
 			enableScripts: true,
 			localResourceRoots: [
-				vscode.Uri.file(path.join(context.extensionPath, 'media')),
+				vscode.Uri.file(mediaPath),
 			],
 			retainContextWhenHidden: true,
 		});
 		
 		panel.iconPath = {
-			dark: vscode.Uri.file(path.join(context.extensionPath, 'media', 'icons', 'icon-dark.svg')),
-			light: vscode.Uri.file(path.join(context.extensionPath, 'media', 'icons', 'icon-light.svg')),
+			dark: vscode.Uri.file(path.join(iconsPath, 'icon-dark.svg')),
+			light: vscode.Uri.file(path.join(iconsPath, 'icon-light.svg')),
 		};
 		
 		const diffPanel = new DiffPanel(panel, context, uris, compare);;